requirements.txt
attrs==23.1.0
blinker==1.7.0
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
dill==0.3.7
docopt==0.6.2
facenet-pytorch==2.5.3
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
google-pasta==0.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pathos==0.3.1
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==3.11.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.25.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
s3transfer==0.7.0
sagemaker==2.196.0
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sympy==1.12
tblib==1.7.0
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
